refactoring - removed 'offsite' element from names tesco lotus uses the same functionality, and tesco is not offsite method, so we need to change names for more general.
