+++ Model/Config/Alipay.php
@@ -1,5 +1,5 @@
 <?php
-namespace Omise\Payment\Model\Config\Offsite;
+namespace Omise\Payment\Model\Config;
 
 use Omise\Payment\Model\Config\Config;
