+++ Observer/InternetbankingDataAssignObserver.php
@@ -5,7 +5,7 @@ use Magento\Framework\Event\Observer;
 use Magento\Payment\Observer\AbstractDataAssignObserver;
 use Magento\Quote\Api\Data\PaymentInterface;
 
-class OffsiteInternetbankingDataAssignObserver extends AbstractDataAssignObserver
+class InternetbankingDataAssignObserver extends AbstractDataAssignObserver
 {
     /**
      * @var string
